Not quite done implementing everything
But the program is coming along nicely. I'm in the middle of accounting for ambiguous
nodes. After that I will have to make sure that everything is being passed to the
kernel, in general and within the node loop. I will then have to make sure all of
the arrays that are being passed were properly constructed. lNodeResolutions->lData,
for instance.