Swapped the way that the model cache is stored and accessed
Previously it was by ParentCharacter then child character. Swapped
it such that all parent characters of a given child character are
grouped. This means that all threads are accesssing the same 64 double
block at once, instead of throughout the cache. Presumably this is a
coalescing move, at the expense of bank conflicts? Would spreading
the threads out over the child character blocks reduce bank conflicts but
keep the coalesced nature of this accession? That is the next step, as
well as investigating new caching possibilities, even superior accession
patterns and implementing scaling to move to float based calculations.