Started working on the entirely new granularity
This might well be quite a bit faster. It uses the "Matrix Multiply"
characteristics of this calculation (which of course this is) to
hopefully speed things up by increasing the number of work items per
work group, decrease the total number of accessions per unit
arithmetic, and provide a bit more of a generic solution.

http://gpgpu-computing4.blogspot.com/2009/10/matrix-multiplication-3-opencl.html

for an optimized example of essentially what I am doing. To be
successful as a biologist in an Bioinformatics world I must strive to
accelerate my work using math, not heuristics. The author of this article
saw a ~300x increase in performance with large arrays (perhaps larger
than we're using), but in our case only some of the memory needs to be
moved onto the GPU before the computation.