Unresolved scaling issues remain
Output suggests that the scalings buffer is either being incorrectly
written to or accessed. What was going wrong with the 1D buffer
remains to be seen. The Scalings buffer should model the node buffer
exactly, and yet the results totally different.

Perhaps the layer of separation between the read and printing imparts
some easily corrected error.

To Be Determined...