Finished parallelizing that last bit of the main
The part that generates the result. The problem was the shared
result and += operand. Including the "pragma opm atomic" directive
worked but was slow. Instead I split it into a tempArray filling
and subsequent parallelized reduction. This is only a bit faster
with this sample size, but performance should scale well.

Bug still outstanding: scaling on NVIDIA hardweare.