Working on os x
But to get it here I had to modify the linker code to include the
OpenCL framework. This needs to detect OS X and be automatic.

Other than that everything seems to work.

Ambiguous node code still does not produce the correct answer (though wouldn't
the answer be different once I change the node?)

To check I need to compare it agains the regular code with an ambigous node.
As the single core version isn't workign this will have to be with the
regular MP2 version.

Also the matrix exponentiation thread count can be increased to use
an appropriate number of threads for the system automatically...