I think ambiguous nodes are working
The results of the code now agree with the current multicore
HYPHY implementation but not with the single core version.

The OpenCL version doesn't work with IntermediateCodon.bf either,
though this may as a result of some other optimization calculation
missing from the OpenCL implementation.

Performance stil isn't great, so it is on to that.