Unresolved but narrowed gap/ambig/scaling problem
It appears that one full site is zeroes in the root_cache.

Confirm using line break at end of each site, then investigate
what could be causing such a problem. Also, some of the sites are
incorrect anyway, so this could be a structural issue (indexing)
rather than an accuracy issue.