Continued analysis turns up interesting results, no solution
First of all all of the buffers are writeable. If you removed
protections on the results buffer, for instance, they will overrun
into the node_cache. Even the Node_cache is writeable. With 1174
nucleotides in the input there are 356 sites and the node_cache
can be arbitrarily filled by the leaf and internal kernels
throughout. If the nucleotide number is increased beyond 1174,
however, the site number increases to 357 etc, and for whatever
reason the node_cache doesn't get fully filled. For instance at
1175 the node_cache isn't filled below 45696, which is 128*357.
This indicates that the first two nodes are either not being
filled (likely, though testing with a sentinel initial value
might be in order) or they are being overwritten. Increasing the
site number and monitoring the empty parts of the node_cache might
also expose the source of such an anomaly.

As it stands the kernel's write lines are messed up for the sake
of profiling, and gDEBugger is incredibly useful.

The only other buffer suffering from these problems is the scaling
cache, which is identical in size and indexing to the node_cache
for obvious reasons.