Narrowed problem, still no solution
The problem is that when the sitecount is increased beyond 1174
the leafnode doesn't do anything. Not only does it not do anything
it consumes no time doing nothing. Not even a nanosecond. So
obviously there is something environmentally wrong. It is likely
not a driver bug because the same flaw exists on my NVIDIA system
at home.

tl;dr: Leaf Kernel does squat when sitecount > 1174. Why?