Getting some very interesting results on the mac.
Invalid command queue errors when trying to write the model over
the first time.

Out of frustration I tried to take my mind off the issue by
targetting the CPU. With 60 work groups, however, it failed to run.
Changing the workgroup count to the nearest power of two and
targetting the GPU results in the same error in the same place,
Invalid command queue when writing the model the first time.

Unsure of whether or not I am now able to target the CPU without
error. Will check later