This branch has been commandeered for the CPU implementation
I am making progress in this regard, however. I have discovered
that the segmentation faults are as a result of a read of the model
cache, but not all reads. Only once I add in the NodeID multiplier
do I have problems.

This could implicate all nodeID's greater than 1, suggesting there
is something wrong with the parts model cache after the first node.

I doubt something is wrong with nodeID, but I suppose that is also
an option.