Still no resolution to the kernel problems.
I based a unified kernel off of the ambig kernel, but of course
it didn't work either. When the internal kernel is reactivated,
however, it works fine, even if it is the only kernel queued and
queued for all nodes. If you replace its arguments with those of
the unified kernel it fails, however.

I am now in the process of adding the arguments back one by one
until it again coughs.

Wish me luck.