Long codon is working and correct!
The problem was that I was using constant memory to store the
node flag cache, and it was running out when the site number
increased. Instead of giving a warning, crashing, or erroring out,
it decided to simply have those kernels that had the constant
memory as an argument do absolutely nothing.

jerks.