So the problem stems from the double/float issue
I think the original final implementation was such that the array
was indexed as a float array on the GPU regardless of buffer type.
This caused obvious problems,

However the value stored in the array was correct, which is odd.
To continue testing you can switch off the buffer type declaration
in the kernel arguments block.