Getting very close to a full solution
By changing the ifdef in the function header of results kernel to
a typedef'd buffer type, the problem was largely resolved. There
remains the problem that the result is wayy too high. This has
been isolated to the first parallel reduction, and needs to be
fixed and tested from there out.

Of course this wont work on any computer using floats, so that
portion of the code needs to be tested as well, on both dev/exec
and float only machines.